requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
appwrite==5.1.1
gunicorn==21.2.0
geopy==2.4.1
python-dateutil==2.8.2